# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/meters": {
      "get": {
        "operationId": "MetersController_getAll",
        "summary": "Получить все счетчики",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starting from 1)",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records per page",
            "schema": {}
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Multicolumn search term",
            "schema": {}
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Limit columns to which apply 'search' term",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Format: _field_:_direction_ [direction may be ASC or DESC] e.g. id:DESC",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MeterEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Счетчики"
        ]
      },
      "put": {
        "operationId": "MetersController_upsert",
        "summary": "Создать или обновить счетчик",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterEntity"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        },
        "tags": [
          "Счетчики"
        ]
      }
    },
    "/api/meters/{serial}": {
      "get": {
        "operationId": "MetersController_getBySerial",
        "summary": "Получить счетчик по серийному номеру",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterEntity"
                }
              }
            }
          },
          "404": {
            "description": "Счетчик не найден"
          }
        },
        "tags": [
          "Счетчики"
        ]
      }
    },
    "/api/meters/{serial}/metrics": {
      "get": {
        "operationId": "MetersController_getMetricsBySerial",
        "summary": "Получить метрики по серийному номеру счетчика",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starting from 1)",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records per page",
            "schema": {}
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Multicolumn search term",
            "schema": {}
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Limit columns to which apply 'search' term",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Format: _field_:_direction_ [direction may be ASC or DESC] e.g. id:DESC",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MetricEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Счетчик не найден"
          }
        },
        "tags": [
          "Счетчики"
        ]
      }
    },
    "/api/meters/{serial}/payments": {
      "get": {
        "operationId": "MetersController_getPaymentsBySerial",
        "summary": "Получить оплаты по серийному номеру счетчика",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starting from 1)",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records per page",
            "schema": {}
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Multicolumn search term",
            "schema": {}
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Limit columns to which apply 'search' term",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Format: _field_:_direction_ [direction may be ASC or DESC] e.g. id:DESC",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PaymentEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Счетчик не найден"
          }
        },
        "tags": [
          "Счетчики"
        ]
      }
    },
    "/api/meters/{serial}/pay": {
      "post": {
        "operationId": "MetersController_payBySerial",
        "summary": "Оплатить счетчик по серийному номеру",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Счетчики"
        ]
      }
    },
    "/api/meters/pay": {
      "post": {
        "operationId": "MetersController_pay",
        "summary": "Оплатить все счетчики",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Счетчики"
        ]
      }
    },
    "/api/metrics": {
      "get": {
        "operationId": "MetricsController_getAll",
        "summary": "Получить все метрики",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starting from 1)",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records per page",
            "schema": {}
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Multicolumn search term",
            "schema": {}
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Limit columns to which apply 'search' term",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Format: _field_:_direction_ [direction may be ASC or DESC] e.g. id:DESC",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MetricEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Метрики"
        ]
      }
    },
    "/api/payments": {
      "get": {
        "operationId": "PaymentsController_getAll",
        "summary": "Получить все оплаты",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starting from 1)",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records per page",
            "schema": {}
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Multicolumn search term",
            "schema": {}
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Limit columns to which apply 'search' term",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Format: _field_:_direction_ [direction may be ASC or DESC] e.g. id:DESC",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PaymentEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Оплаты"
        ]
      }
    }
  },
  "info": {
    "title": "iot.back",
    "description": "IoT project documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "PaginatedResponseMetaDto": {
        "type": "object",
        "properties": {
          "itemsPerPage": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "sortBy": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "searchBy": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "search": {
            "type": "string"
          },
          "filter": {
            "type": "object"
          }
        },
        "required": [
          "itemsPerPage",
          "totalItems",
          "currentPage",
          "totalPages",
          "sortBy",
          "searchBy",
          "search"
        ]
      },
      "PaginatedResponseLinksDto": {
        "type": "object",
        "properties": {
          "first": {
            "type": "string"
          },
          "previous": {
            "type": "string"
          },
          "current": {
            "type": "string"
          },
          "next": {
            "type": "string"
          },
          "last": {
            "type": "string"
          }
        },
        "required": [
          "current"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginatedResponseMetaDto"
          },
          "links": {
            "$ref": "#/components/schemas/PaginatedResponseLinksDto"
          }
        },
        "required": [
          "meta",
          "links"
        ]
      },
      "MeterEntity": {
        "type": "object",
        "properties": {
          "serial": {
            "type": "string",
            "example": "00000001",
            "minLength": 8,
            "maxLength": 8
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "serial",
          "description"
        ]
      },
      "MetricEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "createdAt",
          "value"
        ]
      },
      "PaymentEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "timestamp"
        ]
      },
      "CreateMeterDto": {
        "type": "object",
        "properties": {
          "serial": {
            "type": "string",
            "example": "00000001",
            "minLength": 8,
            "maxLength": 8
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "serial"
        ]
      },
      "CreateMetricDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}